User RunE to return command errors
