+++ cmd/generate.go
@@ -1,11 +1,23 @@
 package cmd
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/spf13/cobra"
 	"github.com/sylviamoss/diceware-cli/diceware"
 )
+
+func init() {
+	generateCmd.Flags().StringVar(&dicewareConfig.Lang, "lang", "en", "password language\n available langs: en, pt")
+	generateCmd.Flags().StringVar(&dicewareConfig.Separator, "separator", "/", "character that separates the words.\nuse --separator=none to remove reparator")
+	generateCmd.Flags().Int32Var(&dicewareConfig.Size, "size", 6, "the amount words the password will have")
+	generateCmd.Flags().BoolVar(&dicewareConfig.Pbcopy, "copy", false, "pbcopy password")
+	generateCmd.Flags().BoolVar(&dicewareConfig.Hide, "hide", false, "pbcopy and hide password. You WON'T see the password")
+	generateCmd.Flags().BoolVar(&dicewareConfig.Lower, "lower", false, "remove capitalized first letters")
+	generateCmd.Flags().BoolVar(&dicewareConfig.RemoveNumber, "remove-number", false, "removes the random number we add by default")
+	rootCmd.AddCommand(generateCmd)
+}
 
 var (
 	dicewareConfig diceware.Config
@@ -15,10 +27,12 @@ var (
 		Short: "Generates a diceware password with custom configuration.",
 		Long: `Generates strong passwords based on easily memorable passwords that are 
 	also extremely resistant to attack.`,
-		Run: func(cmd *cobra.Command, args []string) {
+		RunE: func(cmd *cobra.Command, args []string) error {
 			if err := dicewareConfig.Generate(); err != nil {
-				fmt.Printf("Ops...something went wrong: %s", err.Error())
+				errorMsg := fmt.Sprintf("Ops...something went wrong: %s", err.Error())
+				return errors.New(errorMsg)
 			}
+			return nil
 		},
 	}
 )
